rename: 폴더 구조, 폴더명 변경 (#7)
