+++ src/pages/index.ts
@@ -1,5 +1,5 @@
 export { default as Main } from "./Main";
-export { default as Login } from "./Login/LoginPage";
+export { default as Login } from "./Login";
 export { default as My } from "./My";
 export { default as Planner } from "./Planner";
 export { default as NotFound } from "./NotFound";
